refactor algorithm.const to algorithm.enum in imports
