+++ index.ts
@@ -1,4 +1,4 @@
-export {Algorithm} from "./lib/algorithm.const";
+export {Algorithm} from "./lib/algorithm.enum";
 import {decrypt} from "./lib/decrypt.function";
 import {encrypt} from "./lib/encrypt.function";
 import {generateKey} from "./lib/generate-key";
